Revert "perf(web): defer EVM wallet connectors"
This reverts commit 3cd7645d621a6a7e349cc81924a942d6b1e353fa.
