+++ apps/web/src/lib/wallet/namespaces/evm/adapters/safe.ts
@@ -25,3 +25,18 @@ export const getSafeConnector = async () => {
 
   return await getCreateConnectorFn()
 }
+
+export async function isSafeAppAvailable(): Promise<boolean> {
+  if (typeof window === 'undefined') return false
+
+  try {
+    const { default: SafeAppsSDK } = await import('@safe-global/safe-apps-sdk')
+    const sdk = new SafeAppsSDK()
+
+    // In a Safe iframe this resolves with safe info; outside it typically rejects/throws.
+    await sdk.safe.getInfo()
+    return true
+  } catch {
+    return false
+  }
+}
